+++ openrtb-core/src/main/java/com/google/openrtb/json/OpenRtbNativeJsonReader.java
@@ -20,6 +20,7 @@ import static com.google.openrtb.json.OpenRtbJsonUtils.endArray;
 import static com.google.openrtb.json.OpenRtbJsonUtils.endObject;
 import static com.google.openrtb.json.OpenRtbJsonUtils.getCurrentName;
 import static com.google.openrtb.json.OpenRtbJsonUtils.getIntBoolValue;
+import static com.google.openrtb.json.OpenRtbJsonUtils.peekToken;
 import static com.google.openrtb.json.OpenRtbJsonUtils.startArray;
 import static com.google.openrtb.json.OpenRtbJsonUtils.startObject;
 
@@ -101,7 +102,7 @@ public class OpenRtbNativeJsonReader extends AbstractOpenRtbJsonReader {
       if (par.nextToken() != JsonToken.VALUE_NULL) {
         if (firstField) {
           firstField = false;
-          if ((rootNativeField = "native".equals(fieldName))) {
+          if ((rootNativeField = "native".equals(fieldName)) == true) {
             startObject(par);
             fieldName = getCurrentName(par);
             par.nextToken();
@@ -337,7 +338,7 @@ public class OpenRtbNativeJsonReader extends AbstractOpenRtbJsonReader {
       if (par.nextToken() != JsonToken.VALUE_NULL) {
         if (firstField) {
           firstField = false;
-          if ((rootNativeField = "native".equals(fieldName))) {
+          if ((rootNativeField = "native".equals(fieldName)) == true) {
             startObject(par);
             fieldName = getCurrentName(par);
             par.nextToken();
@@ -492,8 +493,13 @@ public class OpenRtbNativeJsonReader extends AbstractOpenRtbJsonReader {
       throws IOException {
     switch (fieldName) {
       case "vasttag":
-        for (startArray(par); endArray(par); par.nextToken()) {
-          video.addVasttag(par.getText());
+        if (peekToken(par) == JsonToken.START_ARRAY) {
+          // Compatibility with old buggy proto; remove some day.
+          for (startArray(par); endArray(par); par.nextToken()) {
+            video.setVasttag(par.getText());
+          }
+        } else {
+          video.setVasttag(par.getText());
         }
         break;
       default: